added git hub issue #15 reproducer (issue fixed by the previous commit) moved several tests back to the original packages
