+++ tap4j/src/test/java/org/tap4j/parser/TestConstants.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.tap4j.model;
+package org.tap4j.parser;
 
 import static org.junit.Assert.assertNotNull;
 
@@ -29,6 +29,7 @@ import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 
 import org.junit.Test;
+import org.tap4j.model.Patterns;
 
 /**
  * Tests for parser Patterns.
