+++ tap4j/src/test/java/org/tap4j/parser/issueGitHub15/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Tests for GitHub issue #15.
+ * 
+ * subtest as first test causes ParserException (Invalid indentantion. Check your TAP Stream.) by @jochenk
+ */
+package org.tap4j.parser.issueGitHub15;
